# first_shot/openapi.json
{
  "openapi": "3.1.0", 
  "info": {
    "title": "FastAPI", 
    "version": "0.1.0"
  }, 
  "paths": {
    "/": {
      "get": {
        "summary": "Read Root", 
        "operationId": "read_root__get", 
        "responses": {
          "200": {
            "description": "Successful Response", 
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }, 
    "/items/{item_id}": {
      "get": {
        "summary": "Read Item", 
        "operationId": "read_item_items__item_id__get", 
        "parameters": [
          {
            "name": "item_id", 
            "in": "path", 
            "required": true, 
            "schema": {
              "type": "integer", 
              "title": "Item Id"
            }
          }, 
          {
            "name": "q", 
            "in": "query", 
            "required": false, 
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                }, 
                {
                  "type": "null"
                }
              ], 
              "title": "Q"
            }
          }
        ], 
        "responses": {
          "200": {
            "description": "Successful Response", 
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }, 
          "422": {
            "description": "Validation Error", 
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  }, 
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }, 
            "type": "array", 
            "title": "Detail"
          }
        }, 
        "type": "object", 
        "title": "HTTPValidationError"
      }, 
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                }, 
                {
                  "type": "integer"
                }
              ]
            }, 
            "type": "array", 
            "title": "Location"
          }, 
          "msg": {
            "type": "string", 
            "title": "Message"
          }, 
          "type": {
            "type": "string", 
            "title": "Error Type"
          }
        }, 
        "type": "object", 
        "required": [
          "loc", 
          "msg", 
          "type"
        ], 
        "title": "ValidationError"
      }
    }
  }
}